refactor: 💡 move events to config.ts
BREAKING CHANGE: 🧨 None
